+++ web/documentserver-example/nodejs/helpers/users.js
@@ -30,6 +30,7 @@ class User {
     descriptions,
     templates,
     avatar,
+    close,
   ) {
     this.id = id;
     this.name = name;
@@ -43,6 +44,7 @@ class User {
     this.descriptions = descriptions;
     this.templates = templates;
     this.avatar = avatar;
+    this.close = close;
   }
 }
 
@@ -112,6 +114,7 @@ const users = [
     descrUser1,
     true,
     true,
+    { visible: false },
   ),
   new User(
     'uid-2',
@@ -130,6 +133,7 @@ const users = [
     descrUser2,
     false,
     true,
+    {},
   ),
   new User(
     'uid-3',
@@ -148,8 +152,9 @@ const users = [
     descrUser3,
     false,
     false,
+    {},
   ),
-  new User('uid-0', null, null, null, null, {}, [], null, ['protect'], descrUser0, false, false),
+  new User('uid-0', null, null, null, null, {}, [], null, ['protect'], descrUser0, false, false, null),
 ];
 
 // get a list of all the users